implement operations as tabs within request panel
