+++ packages/docusaurus-plugin-openapi-docs/src/markdown/createStatusCodes.ts
@@ -54,6 +54,7 @@ export default function json2xml(o: any, tab: any) {
 }
 
 interface Props {
+  label?: string;
   responses: ApiItem["responses"];
 }
 
@@ -254,7 +255,7 @@ export function createExampleFromSchema(schema: any, mimeType: string) {
   return undefined;
 }
 
-export function createStatusCodes({ responses }: Props) {
+export function createStatusCodes({ label, responses }: Props) {
   if (responses === undefined) {
     return undefined;
   }
@@ -269,6 +270,7 @@ export function createStatusCodes({ responses }: Props) {
       create("div", {
         children: [
           create("ApiTabs", {
+            label,
             children: codes.map((code) => {
               const responseHeaders: any = responses[code].headers;
               return create("TabItem", {
