+++ packages/docusaurus-theme-openapi-docs/src/theme/ApiTabs/index.js
@@ -15,7 +15,14 @@ import useIsBrowser from "@docusaurus/useIsBrowser";
 import Heading from "@theme/Heading";
 import clsx from "clsx";
 
-function TabList({ className, block, selectedValue, selectValue, tabValues }) {
+function TabList({
+  className,
+  block,
+  selectedValue,
+  selectValue,
+  tabValues,
+  label,
+}) {
   const tabRefs = [];
   const { blockElementScrollPositionUntilNextRender } =
     useScrollPositionBlocker();
@@ -85,7 +92,7 @@ function TabList({ className, block, selectedValue, selectValue, tabValues }) {
   return (
     <div className="openapi-tabs__response-header-section">
       <Heading as="h2" id="responses" className="openapi-tabs__response-header">
-        Responses
+        {label || "Responses"}
       </Heading>
       <div className="openapi-tabs__response-container">
         {showTabArrows && (
