--- packages/docusaurus-theme-openapi-docs/src/theme/CallbacksTabs/index.js
@@ -1,192 +1,0 @@
-/* ============================================================================
- * Copyright (c) Palo Alto Networks
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- * ========================================================================== */
-
-import React, { cloneElement, useRef, useState, useEffect } from "react";
-
-import {
-  useScrollPositionBlocker,
-  useTabs,
-} from "@docusaurus/theme-common/internal";
-import useIsBrowser from "@docusaurus/useIsBrowser";
-import Heading from "@theme/Heading";
-import clsx from "clsx";
-
-function TabList({ className, block, selectedValue, selectValue, tabValues }) {
-  const tabRefs = [];
-  const { blockElementScrollPositionUntilNextRender } =
-    useScrollPositionBlocker();
-
-  const handleTabChange = (event) => {
-    const newTab = event.currentTarget;
-    const newTabIndex = tabRefs.indexOf(newTab);
-    const newTabValue = tabValues[newTabIndex].value;
-    if (newTabValue !== selectedValue) {
-      blockElementScrollPositionUntilNextRender(newTab);
-      selectValue(newTabValue);
-    }
-  };
-
-  const handleKeydown = (event) => {
-    let focusElement = null;
-    switch (event.key) {
-      case "Enter": {
-        handleTabChange(event);
-        break;
-      }
-      case "ArrowRight": {
-        const nextTab = tabRefs.indexOf(event.currentTarget) + 1;
-        focusElement = tabRefs[nextTab] ?? tabRefs[0];
-        break;
-      }
-      case "ArrowLeft": {
-        const prevTab = tabRefs.indexOf(event.currentTarget) - 1;
-        focusElement = tabRefs[prevTab] ?? tabRefs[tabRefs.length - 1];
-        break;
-      }
-      default:
-        break;
-    }
-    focusElement?.focus();
-  };
-
-  const tabItemListContainerRef = useRef(null);
-  const [showTabArrows, setShowTabArrows] = useState(false);
-
-  useEffect(() => {
-    const resizeObserver = new ResizeObserver((entries) => {
-      for (let entry of entries) {
-        if (entry.target.offsetWidth < entry.target.scrollWidth) {
-          setShowTabArrows(true);
-        } else {
-          setShowTabArrows(false);
-        }
-      }
-    });
-
-    resizeObserver.observe(tabItemListContainerRef.current);
-
-    return () => {
-      resizeObserver.disconnect();
-    };
-  }, []);
-
-  const handleRightClick = () => {
-    tabItemListContainerRef.current.scrollLeft += 90;
-  };
-
-  const handleLeftClick = () => {
-    tabItemListContainerRef.current.scrollLeft -= 90;
-  };
-
-  return (
-    <div className="callbacks-tabs__callbacks-header-section">
-      <Heading
-        as="h2"
-        id="callbacks"
-        className="callbacks-tabs__callbacks-header"
-      >
-        Callbacks
-      </Heading>
-      <div className="callbacks-tabs__callbacks-container">
-        {showTabArrows && (
-          <button
-            className="callbacks-tabs__arrow left"
-            onClick={handleLeftClick}
-          />
-        )}
-        <ul
-          ref={tabItemListContainerRef}
-          role="tablist"
-          aria-orientation="horizontal"
-          className={clsx(
-            "callbacks-tabs__callbacks-list-container",
-            "tabs",
-            {
-              "tabs--block": block,
-            },
-            className
-          )}
-        >
-          {tabValues.map(({ value, label, attributes }) => (
-            <li
-              // TODO extract TabListItem
-              role="tab"
-              tabIndex={selectedValue === value ? 0 : -1}
-              aria-selected={selectedValue === value}
-              key={value}
-              ref={(tabControl) => tabRefs.push(tabControl)}
-              onKeyDown={handleKeydown}
-              onClick={handleTabChange}
-              {...attributes}
-              className={clsx(
-                "tabs__item",
-                "callbacks-tabs__callbacks-code-item",
-                attributes?.className,
-                "info",
-                {
-                  active: selectedValue === value,
-                }
-              )}
-            >
-              {label ?? value}
-            </li>
-          ))}
-        </ul>
-        {showTabArrows && (
-          <button
-            className="callbacks-tabs__arrow right"
-            onClick={handleRightClick}
-          />
-        )}
-      </div>
-    </div>
-  );
-}
-function TabContent({ lazy, children, selectedValue }) {
-  // eslint-disable-next-line no-param-reassign
-  children = Array.isArray(children) ? children : [children];
-  if (lazy) {
-    const selectedTabItem = children.find(
-      (tabItem) => tabItem.props.value === selectedValue
-    );
-    if (!selectedTabItem) {
-      // fail-safe or fail-fast? not sure what's best here
-      return null;
-    }
-    return cloneElement(selectedTabItem, { className: "margin-top--md" });
-  }
-  return (
-    <div className="margin-top--md">
-      {children.map((tabItem, i) =>
-        cloneElement(tabItem, {
-          key: i,
-          hidden: tabItem.props.value !== selectedValue,
-        })
-      )}
-    </div>
-  );
-}
-function TabsComponent(props) {
-  const tabs = useTabs(props);
-  return (
-    <div className="callbacks-tabs__container">
-      <TabList {...props} {...tabs} />
-      <TabContent {...props} {...tabs} />
-    </div>
-  );
-}
-export default function CallbacksTabs(props) {
-  const isBrowser = useIsBrowser();
-  return (
-    <TabsComponent
-      // Remount tabs after hydration
-      // Temporary fix for https://github.com/facebook/docusaurus/issues/5653
-      key={String(isBrowser)}
-      {...props}
-    />
-  );
-}
