+++ packages/docusaurus-theme-openapi-docs/src/theme/styles.scss
@@ -36,7 +36,7 @@
 @use "./DiscriminatorTabs/DiscriminatorTabs";
 @use "./MimeTabs/MimeTabs";
 @use "./SchemaTabs/SchemaTabs";
-@use "./CallbacksTabs/CallbacksTabs";
+@use "./OperationTabs/OperationTabs";
 /* Code Samples */
 @use "./ResponseSamples/ResponseSamples";
 /* Markdown Styling */
